namkim : branch migration : move storeModule into chatModule
