+++ src/backend/src/chat/chat.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ChatService } from './chat.service';
 import { ChatGateway } from './chat.gateway';
 import { JwtModule } from '@nestjs/jwt';
-import { StoreModule } from 'src/store/store.module';
+import { StoreModule } from 'src/chat/store/store.module';
 
 @Module({
 	imports : [
