+++ src/backend/src/chat/store/store.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { ChatMessageStoreService } from './store.message.service';
 import { ChatUserStoreService } from './store.user.service';
 import { ChatRoomStoreService } from './store.room.service';
-import { GameMatchStoreService } from './store.match.service';
 
 //사용이 좀 이상한 것 같지만...!
 @Module({
@@ -10,7 +9,6 @@ import { GameMatchStoreService } from './store.match.service';
 		ChatMessageStoreService,
 		ChatUserStoreService,
 		ChatRoomStoreService,
-		GameMatchStoreService
 	],
 	exports : [
 		ChatMessageStoreService,
