+++ src/backend/src/chat/store/store.room.service.spec.ts
