+++ src/backend/src/chat/store/store.room.service.ts
