+++ src/backend/src/chat/store/store.user.service.ts
@@ -1,38 +1,31 @@
 import { Injectable } from '@nestjs/common';
 import { Message } from './store.message.service';
 
-//만약 유저가 이름을 바꾼다면.... <- 항상 고려할 사항
 export class User {
-	id : number;	//어차피 map에 있을건데 여기 있을 필요있...? -> 편의!
-	nickname : string;	//as nickname? or will you give me a chance to change nickname?
-		//만약 nick으로 하면 바꿀 때를 구현하기 힘들 것 같다...
-		//아니면 이전까지 기록을 싹 갈아치운 다음에 이걸 가는 것도 가능할 것 같기도?
-	blocklist : Set<number>;	//이건 client Local 에서 관리한다...! 그래도 일단 저장
+	id : number;
+	nickname : string;
+	blocklist : Set<number>;
 	connected : boolean;
 	isGaming : boolean;
-	//들어간 방 모두 가지고 있어야 하나?
-	joinlist : Set<string>; // <-  이것도 이제 필요 없지 않을까? 아니야... 일단 저장하자
+	joinlist : Set<string>;
 	currentRoom : string;
 	messages : Map<number, Message[]>;
-
 	constructor(id : number, nickname : string){
 		this.id = id;
 		this.nickname = nickname;
 		this.blocklist = new Set();
-		this.connected = true;	//접속했을 때 불러오니까!
+		this.connected = true;
 		this.isGaming = false;
-		this.joinlist = new Set();	//default방... 하지만 이건 어차피 enterRoom method로 감쌀 것것 
-		this.messages = new Map();
-		// this.currentRoom = "Default";	//얘도 enterroom이 처리할 것 같다
+		this.joinlist = new Set();
+		// this.messages = new Map();
 	}
 
-	//서버에서 관리가 필요할까?
 	addUserToBlocklist(userid : number){
 		this.blocklist.add(userid);
 	}
-	//서버에서 관리가2
+
 	deleteUserFromBlockList(userid : number){
-		this.blocklist.delete(userid);	//없는 경우 뭔가 조치 필요?
+		this.blocklist.delete(userid);
 	}
 
 	addRoomToJoinlist(roomname : string){
@@ -43,18 +36,18 @@ export class User {
 		this.joinlist.delete(roomname);
 	}
 
-	//getter는 필요없음
+	//메모리 관리 방법을 생각하면 필요없는 것 같은데...
 	private clearSets(){
 		this.blocklist.clear();
 		this.joinlist.clear();
 	}
 
-	//prevent memory leaks ... 맞을까...? chatGPT 신뢰 가능...?
+	//clearSets...?
 	clearUser(){
 		this.clearSets();
 		this.blocklist = null;
 		this.joinlist = null;
-		this.messages = null;	//?
+		// this.messages = null;
 	}
 }
 
@@ -69,11 +62,10 @@ interface UserStore {
 export class ChatUserStoreService implements UserStore{
 	users = new Map();
 	findUserById(id: number): User | undefined {
-		//못 찾으면 undefined
 		return this.users.get(id);
 	}
 
-	//유저가 없으면 만들고 있으면 원래 있는 유저를 반환한다. update는 안됨
+	//create user if it does not exist and return the instance. no update
 	saveUser(id: number, user: User): User {
 		const target = this.users.get(id);
 		if (target === undefined)
@@ -97,7 +89,7 @@ export class ChatUserStoreService implements UserStore{
 			return (user.nickname);
 	}
 
-	//TODO: 이건 안되면 through를 해야하나...?
+	//TODO: 이건 안되면 through를 해야하나...? -> throw하는 방향으로
 	getIdByNickname(nickname : string) : number {
 		const res = this.findAllUser().find((user) => user.nickname === nickname);
 		return (res ? res.id : -1);	//Error_code?	//전역 상수 어떻게 정의?	//number를 undefined로 할 수 있냐구...?ㅠㅠ
