--- src/backend/src/store/store.match.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { GameMatchStoreService } from './store.match.service';
-
-describe('GameMatchStoreService', () => {
-  let service: GameMatchStoreService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [GameMatchStoreService],
-    }).compile();
-
-    service = module.get<GameMatchStoreService>(GameMatchStoreService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
